Mark close abstract; simplify percentile calc
Add @abstractmethod to Executor.close so subclasses are required to implement cleanup logic. In src/benchmarks/cpu_bench.py simplify the percentile index calculation by removing an unnecessary int() cast around round(), since round() yields an integer for this use, making the code clearer.

+++ src/benchmarks/cpu_bench.py
@@ -106,7 +106,7 @@ def _percentile(values: list[float], ratio: float) -> float:
         return 0.0
 
     ordered = sorted(values)
-    index = min(int(round((len(ordered) - 1) * ratio)), len(ordered) - 1)
+    index = min(round((len(ordered) - 1) * ratio), len(ordered) - 1)
     return ordered[index]
